+++ src/components/QrScanner/QrCode.tsx
