+++ src/components/UserInfoSubmission/const.ts
@@ -1,0 +1,13 @@
+export const steps = [
+  { label: "Orderer" },
+  { label: "Company Address" },
+  { label: "Shipping Address" },
+];
+
+export const business = [
+  "Trading",
+  "Wholesaler",
+  "Converter",
+  "Brand",
+  "Student",
+];
