+++ src/components/const.ts
@@ -1,4 +1,19 @@
-import { CountryType } from "./countries";
+export const PRIMARY_DARK = "#323232";
+export const PRIMARY = "#42FF00";
+
+export const META = {
+  title: "YOUNGWON X MAEIL",
+  description: "QR Scan",
+  url: "https://jojoywmaeil.com/",
+  ogImage:
+    "https://server.jojoywmaeil.com:5001/images/v0/b/qrcode-6f8e8.appspot.com/o/logo.png?alt=media&token=fb3a6b7b-0fad-4a37-97e0-d6091babd31b",
+  siteName: "YOUNGWON X MAEIL",
+  type: "website",
+} as const;
+
+export const SERVER_URL = `${process.env.SERVER_URL}`;
+
+import { CountryType } from "@/components/UserInfoSubmission/countries";
 
 interface FormType {
   name: string;
@@ -39,17 +54,3 @@ export const initialValues: FormType = {
   isSameAddress: false,
   productLengthUnit: "METER",
 };
-
-export const steps = [
-  { label: "Orderer" },
-  { label: "Company Address" },
-  { label: "Shipping Address" },
-];
-
-export const business = [
-  "Trading",
-  "Wholesaler",
-  "Converter",
-  "Brand",
-  "Student",
-];
