+++ src/components/pages/MainPage.tsx
@@ -4,9 +4,9 @@ import QrScannerPage from "./QrScannerPage";
 import ToBuyListPage from "./ToBuyListPage";
 import UserInfoSubmissionPage from "./UserInfoSubmissionPage";
 import { BottomAppBar, TitleAppBar } from "../AppBar";
-import { validationSchema } from "@/consts/validation";
-import { initialValues } from "@/consts/form";
-import { SERVER_URL } from "@/consts/url";
+import { validationSchema } from "@/components/validation";
+import { initialValues } from "@/components/const";
+import { SERVER_URL } from "@/components/const";
 import SplashScreen from "../SplashScreen";
 import { useTranslation } from "react-i18next";
 import dayjs from "dayjs";
