+++ src/components/pages/ToBuyListPage.tsx
@@ -1,10 +1,14 @@
 import styled from "styled-components";
 import Icons from "../Icons";
 import { Dispatch, SetStateAction } from "react";
-import Product from "../Product";
+import Product from "../ToBuyList/ToBuyItem";
 import { MessageSnackBar } from "../SnackBar";
 import { Box } from "@mui/material";
-import { SelectedBox, StyledBox, StyledButton } from "../Product/styled";
+import {
+  SelectedBox,
+  StyledBox,
+  StyledButton,
+} from "../ToBuyList/ToBuyItem/styled";
 import { FormikProps } from "formik";
 import { useTranslation } from "react-i18next";
