--- src/consts/colors.tsx
@@ -1,2 +1,0 @@
-export const PRIMARY_DARK = "#323232";
-export const PRIMARY = "#42FF00";
