--- src/consts/metadata.ts
@@ -1,9 +1,0 @@
-export const META = {
-  title: "YOUNGWON X MAEIL",
-  description: "QR Scan",
-  url: "https://jojoywmaeil.com/",
-  ogImage:
-    "https://server.jojoywmaeil.com:5001/images/v0/b/qrcode-6f8e8.appspot.com/o/logo.png?alt=media&token=fb3a6b7b-0fad-4a37-97e0-d6091babd31b",
-  siteName: "YOUNGWON X MAEIL",
-  type: "website",
-} as const;
