--- src/consts/url.ts
@@ -1,1 +1,0 @@
-export const SERVER_URL = `${process.env.SERVER_URL}`;
